requirements.txt
selenium
jupyter
pyautogui
keyboard
python-dotenv